get_map_error now follow the rules
